+++ src/index.js
@@ -7,9 +7,15 @@ if (!observable) throw new Error("mobx-react requires mobx to be available")
 
 if (typeof rdBatched === "function") configure({ reactionScheduler: rdBatched })
 
-export { Observer, useAsObservableSource, useLocalStore } from "mobx-react-lite"
+export {
+    Observer,
+    useAsObservableSource,
+    useLocalStore,
+    isUsingStaticRendering,
+    useStaticRendering
+} from "mobx-react-lite"
 
-export { observer, useStaticRendering } from "./observer"
+export { observer } from "./observer"
 
 export { Provider, MobXProviderContext } from "./Provider"
 export { inject } from "./inject"